pyproject.toml
[project]
name = "nvidia-canary-flash"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10.12"
dependencies = [
    "hf-xet>=1.1.5",
    "nemo-toolkit[asr]>=2.3.1",
]
